+++ spec/models/conversion/project_spec.rb
@@ -4,19 +4,31 @@ RSpec.describe Conversion::Project do
   describe "#route" do
     before { mock_successful_api_responses(urn: any_args, ukprn: any_args) }
 
-    context "when the project is a voluntary conversion" do
-      let(:project) { create(:conversion_project, sponsor_trust_required: false) }
+    context "when a directive academy order has been issued" do
+      context "and the school is joining a sponsor trust" do
+        let(:project) { create(:conversion_project, directive_academy_order: true, sponsor_trust_required: true) }
 
-      it "returns the correct route" do
-        expect(project.route).to eq :voluntary
+        it "the route is sponsored" do
+          expect(project.route).to eq :sponsored
+        end
       end
     end
 
-    context "when the project is a sponsored conversion" do
-      let(:project) { create(:conversion_project, sponsor_trust_required: true) }
+    context "when the project has not been issued a directive academy order" do
+      context "and the school is joining a sponsor trust" do
+        let(:project) { create(:conversion_project, directive_academy_order: false, sponsor_trust_required: true) }
 
-      it "returns the correct route" do
-        expect(project.route).to eq :sponsored
+        it "the route is voluntary" do
+          expect(project.route).to eq :voluntary
+        end
+      end
+
+      context "and the school is not joining a sponsor trust" do
+        let(:project) { create(:conversion_project, directive_academy_order: false, sponsor_trust_required: false) }
+
+        it "the route is voluntary" do
+          expect(project.route).to eq :voluntary
+        end
       end
     end
   end